Fixed bugs and finished Gauss method

+++ Lab2/src/main/kotlin/SeidelMethod.kt
@@ -1,7 +1,5 @@
-import java.math.BigDecimal
-
-class SeidelMethod(val matrix: Array<Array<BigDecimal>>, val vector: Array<BigDecimal>) {
-    val result: Array<BigDecimal> by lazy{
+class SeidelMethod(val matrix: Array<Array<Double>>, val vector: Array<Double>) {
+    val result: Array<Double> by lazy{
         vector
     }
 }
